move csv out of public, change paths
