+++ citizens-guide/csv/mts_mspd_rod_debt.csv
@@ -1,0 +1,12 @@
+year,income,spending,surplus_deficit,interest_public_debt,source_url,gdp,income_gdp,spending_gdp,surplus_deficit_gdp,gdp_url,debt,debt_url,int_rate,int_rate_url
+2001,1990200000000.00,1863040000000.00,127165000000.00,359508000000.00,https://www.fiscal.treasury.gov/fsreports/rpt/mthTreasStmt/mts0901.pdf,1.05963E+13,0.19,0.18,0.01,https://fred.stlouisfed.org/series/GDP,5.80746E+12,ftp://ftp.publicdebt.treas.gov/opd/opds092001.pdf,0.0663,https://www.treasurydirect.gov/govt/rates/pd/avg/2001/2001_09.htm
+2008,2523860000000.00,2978660000000.00,-454806000000.00,451154000000.00,https://www.fiscal.treasury.gov/fsreports/rpt/mthTreasStmt/mts0908.pdf,1.48352E+13,0.17,0.2,-0.03,https://fred.stlouisfed.org/series/GDP,1.00247E+13,ftp://ftp.publicdebt.treas.gov/opd/opds092008.pdf,0.0501,https://www.treasurydirect.gov/govt/rates/pd/avg/2008/2008_09.htm
+2009,2104610000000.00,3521730000000.00,-1417120000000.00,383365000000.00,https://www.fiscal.treasury.gov/fsreports/rpt/mthTreasStmt/mts0909.pdf,1.44203E+13,0.15,0.24,-0.1,https://fred.stlouisfed.org/series/GDP,1.19098E+13,https://treasurydirect.gov/govt/reports/pd/mspd/2009/opds092009.pdf,0.0419,https://www.treasurydirect.gov/govt/rates/pd/avg/2009/2009_09.htm
+2010,2161750000000.00,3455840000000.00,-1294090000000.00,413955000000.00,https://www.fiscal.treasury.gov/fsreports/rpt/mthTreasStmt/mts0910.pdf,1.50799E+13,0.14,0.23,-0.09,https://fred.stlouisfed.org/series/GDP,1.35616E+13,https://treasurydirect.gov/govt/reports/pd/mspd/2010/opds092010.pdf,0.0335,https://www.treasurydirect.gov/govt/rates/pd/avg/2010/2010_09.htm
+2011,2302500000000.00,3601110000000.00,-1298610000000.00,454015000000.00,https://www.fiscal.treasury.gov/fsreports/rpt/mthTreasStmt/mts0911.pdf,1.55919E+13,0.15,0.23,-0.08,https://fred.stlouisfed.org/series/GDP,1.47903E+13,https://treasurydirect.gov/govt/reports/pd/mspd/2011/opds092011.pdf,0.0305,https://www.treasurydirect.gov/govt/rates/pd/avg/2011/2011_09.htm
+2012,2449090000000.00,3538450000000.00,-1089350000000.00,359241000000.00,https://www.fiscal.treasury.gov/fsreports/rpt/mthTreasStmt/mts0912.pdf,1.62572E+13,0.15,0.22,-0.07,https://fred.stlouisfed.org/series/GDP,1.60662E+13,https://treasurydirect.gov/govt/reports/pd/mspd/2012/opds092012.pdf,0.0289,https://www.treasurydirect.gov/govt/rates/pd/avg/2012/2012_09.htm
+2013,2773980000000.00,3454250000000.00,-680276000000.00,415670000000.00,https://www.fiscal.treasury.gov/fsreports/rpt/mthTreasStmt/mts0913.pdf,1.68487E+13,0.16,0.21,-0.04,https://fred.stlouisfed.org/series/GDP,1.67382E+13,https://treasurydirect.gov/govt/reports/pd/mspd/2013/opds092013.pdf,0.0259,https://www.treasurydirect.gov/govt/rates/pd/avg/2013/2013_09.htm
+2014,3020850000000.00,3504200000000.00,-483350000000.00,429568000000.00,https://www.fiscal.treasury.gov/fsreports/rpt/mthTreasStmt/mts0914.pdf,1.77198E+13,0.17,0.2,-0.03,https://fred.stlouisfed.org/series/GDP,1.78241E+13,https://treasurydirect.gov/govt/reports/pd/mspd/2014/opds092014.pdf,0.0243,https://www.treasurydirect.gov/govt/rates/pd/avg/2014/2014_09.htm
+2015,3248720000000.00,3702980000000.00,-454511000000.00,402184000000.00,https://www.fiscal.treasury.gov/fsreports/rpt/mthTreasStmt/mts0915.pdf,1.83311E+13,0.18,0.2,-0.02,https://fred.stlouisfed.org/series/GDP,1.81506E+13,https://treasurydirect.gov/govt/reports/pd/mspd/2015/opds092015.pdf,0.024,https://www.treasurydirect.gov/govt/rates/pd/avg/2015/2015_09.htm
+2016,3266690000000.00,3854100000000.00,-587412000000.00,429963000000.00,https://www.fiscal.treasury.gov/fsreports/rpt/mthTreasStmt/mts0916.pdf,1.87996E+13,0.17,0.21,-0.03,https://fred.stlouisfed.org/series/GDP,1.95734E+13,https://treasurydirect.gov/govt/reports/pd/mspd/2016/opds092016.pdf,0.0235,https://www.treasurydirect.gov/govt/rates/pd/avg/2016/2016_09.htm
+2017,3314890000000.00,3980610000000.00,-665712000000.00,456955000000.00,https://www.fiscal.treasury.gov/fsreports/rpt/mthTreasStmt/mts0917.pdf,1.95881E+13,0.17,0.2,-0.03,https://fred.stlouisfed.org/series/GDP,2.02449E+13,https://treasurydirect.gov/govt/reports/pd/mspd/2017/opds092017.pdf,0.0223,https://www.treasurydirect.gov/govt/rates/pd/avg/2017/2017_09.htm
