+++ citizens-guide/src/income/trends/trendData.js
@@ -1,4 +1,4 @@
-import CategoryData from '../../../public/csv/income_categories_2013_2017.csv';
+import CategoryData from '../../../csv/income_categories_2013_2017.csv';
 
 export function trendData(){
     const indexed = {};
